CRUD_app_Flask/requirement.txt
Flask
Flask-MySQLdb
Flask-Cors
mysql-connector-python
